% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{Maximum Autocorrelation Factors}
\usage{
maf(x)
}
\arguments{
\item{x}{A matrix where each column is a time series. The number of 
rows must be greater than the number of columns}
}
\value{
An object of class "Maf" containing the following items:
\describe{
	\item{\code{x}}{The original data matrix.}
	\item{\code{autocor}}{A vector of autocorrelations for each maf factor.}
 \item{\code{rotation}}{A matrix of rotation vectors, each column representing a vector, 
	with the first column containing the first rotation (MAF1), the second representing
 the second rotation and so on. Each column's squares sum to 1.}
 \item{\code{mafs}}{Contains the matrix with maf factors in the 
	columns, with highest autocorrelation in the first column.}
}
}
\description{
This function performs a linear transformation of the data where the new time series
are mutually orthogonal with identity covariance matrix.
The input is a matrix of column vector time series.
}
\examples{
# Extract mafs from dataset
maf.object = maf(treeringTimeseries)
# Plot the first 6 mafs with undertainty estimates and estimate number of 
# significant mafs contained in the dataset
plot(maf.object) 
}

